update import paths and filenames
