+++ examples/simple_counter_store/src/store.js
@@ -1,3 +1,3 @@
-import createStore from '../../../lif/store.js';
+import createStore from '../../../src/store.js';
 
 export default createStore({ count: 10, type: 'DEC' });
